[OPIK-6833] Extract `OpenAiResponsesStreamingHandler` for response handling
- Moved inline `StreamingChatResponseHandler` implementation to `OpenAiResponsesStreamingHandler` for clarity and reusability.
- Updated `LlmProviderOpenAiResponses` to use the new handler class while preserving existing functionality.
